+++ src/components/Accounts/ViewFeesFines/ViewFeesFines.js
@@ -204,7 +204,7 @@ class ViewFeesFines extends React.Component {
     return {
       '  ': f => (
         <input
-          checked={accounts.find(a => a.id === f.id)}
+          checked={(accounts.findIndex(a => a.id === f.id) !== -1)}
           onClick={e => this.toggleItem(e, f)}
           type="checkbox"
         />
@@ -286,6 +286,12 @@ class ViewFeesFines extends React.Component {
     }));
   }
 
+  rowUpdater = (f) => {
+    const accounts = this.props.selectedAccounts;
+    return this.state.allChecked ||
+    (accounts.findIndex(a => a.id === f.id) !== -1);
+  };
+
   handleOptionsChange(itemMeta, e) {
     e.preventDefault();
     e.stopPropagation();
@@ -347,7 +353,7 @@ class ViewFeesFines extends React.Component {
       <UncontrolledDropdown
         onSelectItem={this.handleOptionsChange}
       >
-        <Button id="ellipsis-button" data-role="toggle" buttonStyle="hover dropdownActive">
+        <Button data-test-ellipsis-button data-role="toggle" buttonStyle="hover dropdownActive">
           <strong>•••</strong>
         </Button>
         <DropdownMenu id="ellipsis-drop-down" data-role="menu">
@@ -439,6 +445,7 @@ class ViewFeesFines extends React.Component {
         sortOrder={sortOrder[0]}
         sortDirection={`${sortDirection[0]}ending`}
         onRowClick={this.onRowClick}
+        rowUpdater={this.rowUpdater}
       />
     );
   }