+++ src/components/Loans/OpenLoans/OpenLoansControl.js
@@ -10,12 +10,11 @@ import {
 
 import { stripesShape } from '@folio/stripes/core';
 
+import { nav } from '../../util';
 import {
-  nav,
-  getOpenRequestsPath,
-} from '../../util';
-
-import { withRenew } from '../../Wrappers';
+  withRenew,
+  withDeclareLost,
+} from '../../Wrappers';
 import TableModel from './components/OpenLoansWithStaticData';
 
 class OpenLoansControl extends React.Component {
@@ -44,6 +43,7 @@ class OpenLoansControl extends React.Component {
       id: PropTypes.string.isRequired,
     }).isRequired,
     history: PropTypes.object.isRequired,
+    location: PropTypes.object.isRequired,
     match: PropTypes.object,
     patronGroup: PropTypes.object.isRequired,
     requestCounts: PropTypes.object.isRequired,
@@ -51,6 +51,7 @@ class OpenLoansControl extends React.Component {
     loans: PropTypes.arrayOf(PropTypes.object).isRequired,
     patronBlocks: PropTypes.arrayOf(PropTypes.object).isRequired,
     renew: PropTypes.func.isRequired,
+    declareLost: PropTypes.func.isRequired,
   };
 
   constructor(props) {
@@ -193,14 +194,7 @@ class OpenLoansControl extends React.Component {
     });
   };
 
-  /**
-   * change handler for the options-menu prevents the event from bubbling
-   * up to the event handler attached to the row.
-   */
-  handleOptionsChange = (itemMeta, e) => {
-    e.preventDefault();
-    e.stopPropagation();
-
+  handleOptionsChange = itemMeta => {
     const {
       loan,
       action,
@@ -209,81 +203,6 @@ class OpenLoansControl extends React.Component {
     if (action && this[action]) {
       this[action](loan);
     }
-  };
-
-  itemDetails(loan, e) {
-    if (e) e.preventDefault();
-
-    const {
-      resources: {
-        query,
-      },
-      mutator: {
-        query: {
-          update,
-        },
-      },
-    } = this.props;
-    const {
-      item:{
-        instanceId,
-        holdingsRecordId,
-      },
-      itemId,
-    } = loan;
-
-    // none of the query params relevent to finding a user
-    // are relevent to finding instances so we purge them all.
-    const q = {};
-    Object.keys(query).forEach((k) => { q[k] = null; });
-
-    update({
-      _path: `/inventory/view/${instanceId}/${holdingsRecordId}/${itemId}`,
-      ...q,
-    });
-  }
-
-  changeDueDate = (loan) => {
-    this.setState({
-      activeLoan: loan.id,
-      changeDueDateDialogOpen: true,
-    });
-  };
-
-  showLoanPolicy = (loan, e) => {
-    if (e) e.preventDefault();
-
-    const {
-      resources: {
-        query,
-      },
-      mutator: {
-        query: {
-          update,
-        },
-      },
-    } = this.props;
-    const q = {};
-
-    Object.keys(query).forEach((k) => { q[k] = null; });
-
-    update({
-      _path: `/settings/circulation/loan-policies/${loan.loanPolicyId}`,
-      ...q,
-    });
-  };
-
-  discoverRequests = (loan) => {
-    const { history } = this.props;
-    const query = get(loan, ['item', 'barcode']);
-    const path = getOpenRequestsPath(query);
-
-    history.push(path);
-  };
-
-  feefine = (loan, e) => {
-    const { history, match: { params } } = this.props;
-    nav.onClickChargeFineToLoan(e, loan, history, params);
   };
 
   renew = (loan) => {
@@ -301,7 +220,16 @@ class OpenLoansControl extends React.Component {
     }
   };
 
-  feefinedetails = (loan, e) => {
+  changeDueDate = loan => {
+    this.setState({
+      activeLoan: loan.id,
+      changeDueDateDialogOpen: true,
+    });
+  };
+
+  declareLost = loan => { this.props.declareLost(loan); };
+
+  feefineDetails = (loan, e) => {
     const {
       resources,
       history,
@@ -372,7 +300,9 @@ class OpenLoansControl extends React.Component {
       patronGroup,
       patronBlocks,
       resources,
+      location,
     } = this.props;
+
     return (
       <div data-test-open-loans>
         <TableModel
@@ -391,6 +321,7 @@ class OpenLoansControl extends React.Component {
           stripes={stripes}
           feeFineCount={this.feeFineCount}
           history={history}
+          location={location}
           match={match}
           user={user}
           toggleAll={this.toggleAll}
@@ -412,4 +343,4 @@ class OpenLoansControl extends React.Component {
   }
 }
 
-export default withRenew(OpenLoansControl);
+export default withDeclareLost(withRenew(OpenLoansControl));