+++ src/components/Loans/OpenLoans/components/ActionsDropdown/ActionsDropdown.js
@@ -2,18 +2,23 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { withRouter } from 'react-router-dom';
+import { get } from 'lodash';
+
 import {
   Button,
-  MenuItem,
   IconButton,
+  Dropdown,
+  DropdownMenu,
 } from '@folio/stripes/components';
 import {
   stripesShape,
   IfPermission,
 } from '@folio/stripes/core';
 
-import Popdown from './Popdown';
-import css from './ActionsDropdown.css';
+import {
+  getChargeFineToLoanPath,
+  getOpenRequestsPath,
+} from '../../../../util';
 
 class ActionsDropdown extends React.Component {
   static propTypes = {
@@ -22,144 +27,116 @@ class ActionsDropdown extends React.Component {
     requestQueue: PropTypes.bool.isRequired,
     handleOptionsChange: PropTypes.func.isRequired,
     disableFeeFineDetails: PropTypes.bool,
-    match: PropTypes.object,
   };
 
-  itemClick = () => {
-    this.props.handleOptionsChange();
-  }
-
-  renderDropdownTrigger = (triggerRef, onToggle, aria) => (
-    <IconButton icon="ellipsis" ref={triggerRef} onClick={onToggle} {...aria} />
-  );
-
-  render() {
+  renderMenu = ({ onToggle }) => {
     const {
       loan,
       handleOptionsChange,
       requestQueue,
       stripes,
       disableFeeFineDetails,
+      match: { params },
     } = this.props;
 
     const itemDetailsLink = `/inventory/view/${loan.item.instanceId}/${loan.item.holdingsRecordId}/${loan.itemId}`;
     const loanPolicyLink = `/settings/circulation/loan-policies/${loan.loanPolicyId}`;
-    const buttonDisabled = !this.props.stripes.hasPerm('ui-users.feesfines.actions.all');
+    const buttonDisabled = !stripes.hasPerm('ui-users.feesfines.actions.all');
 
     return (
-      <Popdown
-        renderTrigger={this.renderDropdownTrigger}
-        portal
-      >
-        <div className={css.actionDropDown}>
-          <IfPermission perm="inventory.items.item.get">
-            <MenuItem
-              itemMeta={{
-                loan,
-                action: 'itemDetails',
-              }}
-              onSelectItem={handleOptionsChange}
-            >
-              <Button
-                buttonStyle="dropdownItem"
-                to={itemDetailsLink}
-                onClick={(e) => { e.stopPropagation(); }}
-              >
-                <FormattedMessage id="ui-users.itemDetails" />
-              </Button>
-            </MenuItem>
-          </IfPermission>
-          <IfPermission perm="ui-users.loans.renew">
-            <MenuItem
-              itemMeta={{
-                loan,
-                action: 'renew',
-              }}
-              onSelectItem={handleOptionsChange}
-            >
-              <Button
-                buttonStyle="dropdownItem"
-                data-test-dropdown-content-renew-button
-                onClick={(e) => {
-                  handleOptionsChange({ loan, action: 'renew' }, e);
-                }}
-              >
-                <FormattedMessage id="ui-users.renew" />
-              </Button>
-            </MenuItem>
-          </IfPermission>
-          <IfPermission perm="ui-users.loans.edit">
-            <MenuItem
-              itemMeta={{
-                loan,
-                action: 'changeDueDate',
-              }}
-              onSelectItem={handleOptionsChange}
-            >
-              <Button
-                buttonStyle="dropdownItem"
-                onClick={(e) => { handleOptionsChange({ loan, action:'changeDueDate' }, e); }}
-                data-test-dropdown-content-change-due-date-button
-              >
-                <FormattedMessage id="stripes-smart-components.cddd.changeDueDate" />
-              </Button>
-            </MenuItem>
-          </IfPermission>
-          <IfPermission perm="circulation-storage.loan-policies.item.get">
-            <MenuItem
-              itemMeta={{
-                loan,
-                action: 'showLoanPolicy',
-              }}
-              onSelectItem={handleOptionsChange}
-            >
-              <Button
-                buttonStyle="dropdownItem"
-                href={loanPolicyLink}
-              >
-                <FormattedMessage id="ui-users.loans.details.loanPolicy" />
-              </Button>
-            </MenuItem>
-          </IfPermission>
-          <MenuItem
-            itemMeta={{
-              loan,
-              action: 'feefine',
-            }}
-            onSelectItem={handleOptionsChange}
+      <DropdownMenu data-role="menu">
+        <IfPermission perm="inventory.items.item.get">
+          <Button
+            buttonStyle="dropdownItem"
+            to={itemDetailsLink}
           >
-            <Button buttonStyle="dropdownItem" disabled={buttonDisabled}>
-              <FormattedMessage id="ui-users.loans.newFeeFine" />
-            </Button>
-          </MenuItem>
-          <MenuItem
-            itemMeta={{
-              loan,
-              action: 'feefinedetails',
+            <FormattedMessage id="ui-users.itemDetails" />
+          </Button>
+        </IfPermission>
+        <IfPermission perm="ui-users.loans.renew">
+          <Button
+            buttonStyle="dropdownItem"
+            data-test-dropdown-content-renew-button
+            onClick={(e) => {
+              handleOptionsChange({ loan, action: 'renew' });
+              onToggle(e);
             }}
-            onSelectItem={handleOptionsChange}
           >
-            <Button buttonStyle="dropdownItem" disabled={disableFeeFineDetails}>
-              <FormattedMessage id="ui-users.loans.feeFineDetails" />
-            </Button>
-          </MenuItem>
-          { requestQueue && stripes.hasPerm('ui-requests.all') &&
-            <MenuItem
-              itemMeta={{
-                loan,
-                action: 'discoverRequests',
-              }}
-              onSelectItem={handleOptionsChange}
-            >
-              <div data-test-dropdown-content-request-queue>
-                <Button buttonStyle="dropdownItem">
-                  <FormattedMessage id="ui-users.loans.details.requestQueue" />
-                </Button>
-              </div>
-            </MenuItem>
-          }
-        </div>
-      </Popdown>
+            <FormattedMessage id="ui-users.renew" />
+          </Button>
+        </IfPermission>
+        <IfPermission perm="ui-users.loans.edit">
+          <Button
+            buttonStyle="dropdownItem"
+            data-test-dropdown-content-change-due-date-button
+            onClick={(e) => {
+              handleOptionsChange({ loan, action:'changeDueDate' });
+              onToggle(e);
+            }}
+          >
+            <FormattedMessage id="stripes-smart-components.cddd.changeDueDate" />
+          </Button>
+        </IfPermission>
+        { get(loan, ['item', 'status', 'name']) !== 'Declared lost' &&
+          <Button
+            buttonStyle="dropdownItem"
+            data-test-dropdown-content-declare-lost-button
+            onClick={(e) => {
+              handleOptionsChange({ loan, action:'declareLost' });
+              onToggle(e);
+            }}
+          >
+            <FormattedMessage id="ui-users.loans.declareLost" />
+          </Button>
+        }
+        <IfPermission perm="circulation-storage.loan-policies.item.get">
+          <Button
+            buttonStyle="dropdownItem"
+            to={loanPolicyLink}
+          >
+            <FormattedMessage id="ui-users.loans.details.loanPolicy" />
+          </Button>
+        </IfPermission>
+        <Button
+          buttonStyle="dropdownItem"
+          disabled={buttonDisabled}
+          to={getChargeFineToLoanPath(params.id, loan.id)}
+        >
+          <FormattedMessage id="ui-users.loans.newFeeFine" />
+        </Button>
+        <Button
+          buttonStyle="dropdownItem"
+          disabled={disableFeeFineDetails}
+          onClick={() => {
+            handleOptionsChange({ loan, action: 'feefineDetails' });
+          }}
+        >
+          <FormattedMessage id="ui-users.loans.feeFineDetails" />
+        </Button>
+        {requestQueue && stripes.hasPerm('ui-requests.all') &&
+          <Button
+            buttonStyle="dropdownItem"
+            data-test-dropdown-content-request-queue
+            to={getOpenRequestsPath(get(loan, ['item', 'barcode']))}
+          >
+            <FormattedMessage id="ui-users.loans.details.requestQueue" />
+          </Button>
+        }
+      </DropdownMenu>
+    );
+  };
+
+  render() {
+    return (
+      <Dropdown
+        renderTrigger={({ getTriggerProps }) => (
+          <IconButton
+            {...getTriggerProps()}
+            icon="ellipsis"
+          />
+        )}
+        renderMenu={this.renderMenu}
+      />
     );
   }
 }