+++ src/index.js
@@ -245,7 +245,7 @@ class UsersRouting extends Component {
               path={`${base}/:id/accounts/:accountstatus/charge`}
               exact
               render={(props) => (
-                <IfPermission perm="ui-users.accounts">
+                <IfPermission perm="ui-users.feesfines.actions.all">
                   <Routes.FeesFinesContainer {...props} />
                 </IfPermission>
               )}
@@ -253,7 +253,7 @@ class UsersRouting extends Component {
             <Route
               path={`${base}/:id/accounts/view/:accountid`}
               render={(props) => (
-                <IfPermission perm="ui-users.accounts">
+                <IfPermission perm="ui-users.feesfines.actions.all">
                   <Routes.AccountDetailsContainer {...props} />
                 </IfPermission>
               )}
@@ -263,7 +263,7 @@ class UsersRouting extends Component {
               exact
               component={Routes.AccountsListingContainer}
               render={(props) => (
-                <IfPermission perm="ui-users.accounts">
+                <IfPermission perm="ui-users.feesfines.actions.all">
                   <Routes.AccountsListingContainer {...props} />
                 </IfPermission>
               )}