+++ src/views/LoanDetails/LoanDetails.js
@@ -25,6 +25,7 @@ import {
   KeyValue,
   Row,
   Col,
+  NoValue,
 } from '@folio/stripes/components';
 import { IfPermission } from '@folio/stripes/core';
 
@@ -34,7 +35,10 @@ import {
   nav,
   getOpenRequestsPath,
 } from '../../components/util';
-import { withRenew } from '../../components/Wrappers';
+import {
+  withRenew,
+  withDeclareLost,
+} from '../../components/Wrappers';
 import loanActionMap from '../../components/data/static/loanActionMap';
 import LoanProxyDetails from './LoanProxyDetails';
 import ViewLoading from '../../components/Loading/ViewLoading';
@@ -66,10 +70,12 @@ class LoanDetails extends React.Component {
     loanPolicies: PropTypes.object,
     requestCounts: PropTypes.object,
     renew: PropTypes.func,
+    declareLost: PropTypes.func,
     patronBlocks: PropTypes.arrayOf(PropTypes.object),
     intl: intlShape.isRequired,
     match: PropTypes.object.isRequired,
     history: PropTypes.object.isRequired,
+    location: PropTypes.object.isRequired,
   };
 
   constructor(props) {
@@ -112,10 +118,7 @@ class LoanDetails extends React.Component {
   }
 
   hideChangeDueDateDialog() {
-    this.setState({
-      changeDueDateDialogOpen: false,
-    });
-    this.props.mutator.modified.replace({ time: new Date().getTime() });
+    this.setState({ changeDueDateDialogOpen: false });
   }
 
   hideNonRenewedLoansModal() {
@@ -251,18 +254,28 @@ class LoanDetails extends React.Component {
   }
 
   handleClose = () => {
-    const { loan, history, match: { params } } = this.props;
+    const {
+      loan,
+      history,
+      match: { params },
+      location,
+    } = this.props;
+
     // if this loan detail was accessed through a fee/fine, accountid will be present.
     if (params.accountid) {
       history.push(`/users/${params.id}/accounts/view/${params.accountid}`);
     }
+
     const loanStatus = loan.status ? loan.status.name.toLowerCase() : 'open';
-    history.push(`/users/${params.id}/loans/${loanStatus}`);
+
+    history.push({
+      pathname: `/users/${params.id}/loans/${loanStatus}`,
+      state: location.state,
+    });
   }
 
   render() {
     const {
-      history,
       loan,
       patronGroup,
       patronBlocks,
@@ -272,6 +285,7 @@ class LoanDetails extends React.Component {
       intl,
       loanPolicies,
       requestCounts,
+      declareLost,
     } = this.props;
 
     const {
@@ -314,6 +328,15 @@ class LoanDetails extends React.Component {
     const contributorsListString = contributorsList.join(' ');
     const contributorsLength = contributorsListString.length;
     const loanStatus = get(loan, ['status', 'name'], '-');
+    const itemStatus = get(loan, ['item', 'status', 'name'], '-');
+    const isDeclaredLostItem = itemStatus === 'Declared lost';
+    let lostDate;
+    const declaredLostActions = loanActionsWithUser.filter(currentAction => get(currentAction, ['action'], '') === 'declaredLost');
+
+    if (isDeclaredLostItem && declaredLostActions.length) {
+      lostDate = get(declaredLostActions[0], ['metadata', 'updatedDate']);
+    }
+
     const buttonDisabled = (loanStatus && loanStatus === 'Closed');
     // Number of characters to truncate the string = 77
     const listTodisplay = (contributorsList === '-') ? '-' : (contributorsListString.length >= 77) ? `${contributorsListString.substring(0, 77)}...` : `${contributorsListString.substring(0, contributorsListString.length - 2)}`;
@@ -338,7 +361,7 @@ class LoanDetails extends React.Component {
             id="pane-loandetails"
             defaultWidth="100%"
             dismissible
-            onClose={() => { history.goBack(); }}
+            onClose={this.handleClose}
             paneTitle={(
               <FormattedMessage id="ui-users.loans.loanDetails">
                 {(loanDetails) => `${loanDetails} - ${getFullName(user)} (${upperFirst(patronGroup.group)})`}
@@ -365,6 +388,14 @@ class LoanDetails extends React.Component {
                     <FormattedMessage id="stripes-smart-components.cddd.changeDueDate" />
                   </Button>
                 </IfPermission>
+                <Button
+                  data-test-declare-lost-button
+                  disabled={buttonDisabled || isDeclaredLostItem}
+                  buttonStyle="primary"
+                  onClick={() => declareLost(loan)}
+                >
+                  <FormattedMessage id="ui-users.loans.declareLost" />
+                </Button>
               </span>
             </Row>
             <Row>
@@ -407,10 +438,13 @@ class LoanDetails extends React.Component {
               </Col>
             </Row>
             <Row>
-              <Col xs={2}>
+              <Col
+                data-test-loan-actions-history-item-status
+                xs={2}
+              >
                 <KeyValue
                   label={<FormattedMessage id="ui-users.loans.columns.itemStatus" />}
-                  value={get(loan, ['item', 'status', 'name'], '-')}
+                  value={itemStatus}
                 />
               </Col>
               <Col xs={2}>
@@ -473,10 +507,13 @@ class LoanDetails extends React.Component {
                   value={requestQueueValue}
                 />
               </Col>
-              <Col xs={2}>
+              <Col
+                data-test-loan-actions-history-lost
+                xs={2}
+              >
                 <KeyValue
                   label={<FormattedMessage id="ui-users.loans.details.lost" />}
-                  value="TODO"
+                  value={lostDate ? (<FormattedTime value={lostDate} day="numeric" month="numeric" year="numeric" />) : (<NoValue />)}
                 />
               </Col>
             </Row>
@@ -523,4 +560,5 @@ class LoanDetails extends React.Component {
 export default compose(
   injectIntl,
   withRenew,
+  withDeclareLost,
 )(LoanDetails);